Copy complete plan file
